working on the relations
